Raise voltage threshold for SBU -> SUB switch
At the end of the day, I end up ping-ponging the battery while
the PV voltage is still high, but there's not enough juice to
power heavy-ish loads. So I'd rather that the unit just back off
and leave itself in SUB mode, especially for those days where
the load shedding comes on late in the day, and you want to make
sure that you're fully charged when that strikes.