Make inverter queries in-process instead of popen()
The USB port on my Axpert VMIII 3KW died, so I had to switch
to RS232. This has the side effect of opening the com port,
talking, and close it, take about 1.5 seconds. By moving this
in-process, and keeping the port open, the query goes down to
about 0.07 seconds. It's also a simpler architecture.
But I should probably add something in here that closes the
file and re-opens it when a command fails. I'll see how this
goes, and if that's necessary.