Raise overload threshold back to 2950
With 'bypass' mode, this inverter can actually handle more
than it's rated power, so I'm no longer worried about blowing
it up by running it close to max, when in hybrid mode.

We still have our MaxLoadBatteryModeW = 1500, which is where
we switch from SBU to SUB (aka off-grid to hybrid), so this
is the safeguard that I want in order to avoid running the inverter
at max capacity.