Make SBU/SUB switching optional, reduce switchover pause
The switchover pause mentioned here that is reduced, is
the not the SBU/SUB switchover pause. The pause that I have
reduced is the switchover between grid and inverter, on
the contactors.
I think 20 was quite conservative, an my aircon would often
switch off when the power was toggled between grid and inverter,
so I'm pretty sure this needs to be reduced.