Introduce hard and soft charge limits
When soft limit is breached, we switch loads from
battery to grid, but charge only from solar.

When hard limit is breached, we switch loads to grid,
and we charge from grid + solar.

If neither of those limits are breached, we run
loads from battery, and charge only from solar.

Also, allow 2 or more goal points to be specified,
and linearly interpolate between them.