Force hard limit <= soft limit
This should smooth out our transitions a bit.
In particular, if you've switched into charge mode, and then you
hit your hard goal + 10, then we'll switch off grid charging,
but still remain in SUB mode until we've also hit soft goal + 10.

This is a small and subtle thing, but I noticed it at the end
of the day when we switched out of grid charge mode, but it would
have been better to remain in SUB mode too.